Move over to dyanmic plugin architecture.
We want to be able to queue up different sequences of animations, scheduled to happen at different times of the day, or when different events occur.
In order to facilitiate this, we need consistency in the way that modules get loaded and configured.
This commit renames the animations folder to plugins, and implements loading the plugins using python importlib.
Using importlib means that we can store the plugins in a list object where we can access them more conveniently.
With this commit, there are no default animations configure at launch and the only way to add them is through the CLI
This commit also includes the frighteningly insecure exec-any-code-you-want cli command which is great for dev, but not for security - will be removed when this becomes a daemon.
New CLI commands created for adding and deleting animations from the queue.